add test:e2e command and add more tests with failing example for atom-input
