+++ package.json
@@ -13,13 +13,14 @@
     "test:aliases": "mocha tests/aliases-test.js",
     "test:core": "mocha tests/core/**/*.js",
     "test:dependencies": "mocha tests/dependencies-test.js",
+    "test:e2e": "mocha tests/e2e/**/*.js",
     "test:examples": "mocha tests/examples-test.js",
     "test:identifiers": "mocha tests/identifier-test.js",
     "test:languages": "mocha tests/run.js",
     "test:patterns": "mocha tests/pattern-tests.js",
     "test:plugins": "mocha tests/plugins/**/*.js",
     "test:runner": "mocha tests/testrunner-tests.js",
-    "test": "npm run test:runner && npm run test:core && npm run test:dependencies && npm run test:languages && npm run test:plugins && npm run test:aliases && npm run test:patterns && npm run test:examples && npm run test:identifiers"
+    "test": "npm run test:runner && npm run test:core && npm run test:dependencies && npm run test:languages && npm run test:plugins && npm run test:aliases && npm run test:patterns && npm run test:examples && npm run test:identifiers && npm run test:e2e"
   },
   "repository": {
     "type": "git",
